Move utils, webcomponents and styles into seperate folders
