+++ openfish-webapp/src/styles/reset.css
@@ -1,0 +1,3 @@
+*, *:before, *:after {
+    box-sizing: border-box;
+}
