+++ openfish-webapp/src/utils/api.types.ts
