+++ openfish-webapp/src/utils/array-utils.ts
