+++ openfish-webapp/src/webcomponents/bounding-box-creator.ts
