+++ openfish-webapp/streams.html
@@ -5,11 +5,11 @@
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>OpenFish</title>
-    <link rel="stylesheet" href="./src/index.css" />
+    <link rel="stylesheet" href="./src/styles/index.css" />
 
-    <script type="module" src="./src/stream-list.ts"></script>
-    <script type="module" src="./src/stream-filter.ts"></script>
-    <script type="module" src="./src/capture-source-dropdown.ts"></script>
+    <script type="module" src="./src/webcomponents/stream-list.ts"></script>
+    <script type="module" src="./src/webcomponents/stream-filter.ts"></script>
+    <script type="module" src="./src/webcomponents/capture-source-dropdown.ts"></script>
 
     <style>
       stream-filter {
