+++ openfish-webapp/test/datetime.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest'
-import { datetimeDifference } from '../src/datetime'
+import { datetimeDifference } from '../src/utils/datetime'
 
 test('datetimeDifference calculates time correctly', () => {
   const a = '2023-12-06T12:17:00.000Z'
