+++ openfish-webapp/watch.html
@@ -5,8 +5,8 @@
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>OpenFish | Watch</title>
-    <link rel="stylesheet" href="./src/index.css" />
-    <script type="module" src="/src/watch-stream.ts"></script>
+    <link rel="stylesheet" href="./src/styles/index.css" />
+    <script type="module" src="/src/webcomponents/watch-stream.ts"></script>
 
     <style>
       main {
